chore(smart-apply): implement the smart-apply provider
