+++ lib/shared/src/index.ts
@@ -63,6 +63,7 @@ export { serializeChatMessage } from './chat/transcript'
 export {
     errorToChatError,
     DEFAULT_EVENT_SOURCE,
+    EventSourceTelemetryMetadataMapping,
 } from './chat/transcript/messages'
 export type {
     AccountKeyedChatHistory,
