+++ vscode/src/edit/provider.ts
@@ -1,9 +1,12 @@
+import { LRUCache } from 'lru-cache'
 import { workspace } from 'vscode'
 import { Utils } from 'vscode-uri'
 
 import {
     BotResponseMultiplexer,
     type CompletionParameters,
+    DEFAULT_EVENT_SOURCE,
+    EventSourceTelemetryMetadataMapping,
     Typewriter,
     currentAuthStatus,
     currentSiteVersion,
@@ -18,22 +21,17 @@ import {
     wrapInActiveSpan,
 } from '@sourcegraph/cody-shared'
 
-import type { FixupController } from '../non-stop/FixupController'
-import type { FixupTask } from '../non-stop/FixupTask'
-import { logError } from '../output-channel-logger'
-
-import {
-    DEFAULT_EVENT_SOURCE,
-    EventSourceTelemetryMetadataMapping,
-} from '@sourcegraph/cody-shared/src/chat/transcript/messages'
 import { doesFileExist } from '../commands/utils/workspace-files'
 import { getEditor } from '../editor/active-editor'
+import type { FixupController } from '../non-stop/FixupController'
+import type { FixupTask } from '../non-stop/FixupTask'
 import { CodyTaskState } from '../non-stop/state'
+import { logError } from '../output-channel-logger'
 import { splitSafeMetadata } from '../services/telemetry-v2'
 import { countCode } from '../services/utils/code-count'
 import { resolveRelativeOrAbsoluteUri } from '../services/utils/edit-create-file'
-import type { EditCacheManager } from './cache-manager'
-import type { EditManagerOptions } from './manager'
+
+import type { EditManagerOptions } from './edit-manager'
 import { responseTransformer } from './output/response-transformer'
 import { buildInteraction } from './prompt'
 import { PROMPT_TOPICS } from './prompt/constants'
@@ -42,8 +40,7 @@ import { isStreamedIntent } from './utils/edit-intent'
 
 interface EditProviderOptions extends EditManagerOptions {
     task: FixupTask
-    controller: FixupController
-    cacheManager: EditCacheManager
+    fixupController: FixupController
 }
 
 /**
@@ -77,20 +74,22 @@ export class EditProvider {
     private insertionQueue: { response: string; isMessageInProgress: boolean }[] = []
     private insertionInProgress = false
 
+    private cache = new LRUCache<string, StreamSession>({ max: 20 })
+
     constructor(public config: EditProviderOptions) {}
 
     /**
      * Attempt to start streaming in "prefetch mode." If a stream is already
      * in progress for this task, do nothing.
      */
-    public prefetchEdit(): void {
+    public prefetchStreamingEdit(): void {
         const { task } = this.config
-        const session = this.config.cacheManager.getStreamSession(task.id)
+        const session = this.cache.get(task.id)
         if (!session) {
             // Kick off streaming in the background (but do not mark the task as started)
             this.performStreamingEdit({ taskId: task.id, initialState: 'prefetching' }).catch(error => {
                 // Remove the broken session so subsequent tries can retry
-                this.config.cacheManager.delete(task.id)
+                this.cache.delete(task.id)
                 // Rethrow so logs are visible
                 throw error
             })
@@ -103,9 +102,9 @@ export class EditProvider {
      * reuse the partial text. If no streaming session exists yet, we start it
      * normally. If streaming completed, we apply the final text instantly.
      */
-    public async startEdit(): Promise<void> {
+    public async startStreamingEdit(): Promise<void> {
         const taskId = this.config.task.id
-        const session = this.config.cacheManager.getStreamSession(taskId)
+        const session = this.cache.get(taskId)
 
         // If no streaming session or there was an error, start from scratch
         if (!session) {
@@ -116,14 +115,14 @@ export class EditProvider {
         // If streaming is already complete, just apply the final partial text
         if (session.state === 'completed') {
             // Mark the task started for UI
-            this.config.controller.startTask(this.config.task)
+            this.config.fixupController.startTask(this.config.task)
             // The final text is in session.partialText
             return this.handleResponse(session.partialText, false)
         }
 
         // If streaming is still in progress, we "startTask" now for UI,
         // then replay what has arrived so far, and continue to stream new tokens.
-        this.config.controller.startTask(this.config.task)
+        this.config.fixupController.startTask(this.config.task)
         session.state = 'streaming'
         // Immediately apply what has already been streamed
         if (session.partialText) {
@@ -137,7 +136,7 @@ export class EditProvider {
 
     public abortEdit(): void {
         const taskId = this.config.task.id
-        const session = this.config.cacheManager.getStreamSession(taskId)
+        const session = this.cache.get(taskId)
         if (session) {
             session.abortController.abort()
         }
@@ -148,7 +147,7 @@ export class EditProvider {
      * We still call "startTask" first to ensure the UI updates accordingly.
      */
     public applyEdit(response: string): Promise<void> {
-        this.config.controller.startTask(this.config.task)
+        this.config.fixupController.startTask(this.config.task)
         return this.handleResponse(response, false)
     }
 
@@ -170,13 +169,13 @@ export class EditProvider {
             streamingPromise: null,
             partialText: '',
         }
-        this.config.cacheManager.setStreamSession(taskId, session)
+        this.cache.set(taskId, session)
 
         session.streamingPromise = wrapInActiveSpan('command.edit.streaming', async span => {
             span.setAttribute('sampled', true)
 
             if (session.state === 'streaming') {
-                this.config.controller.startTask(this.config.task)
+                this.config.fixupController.startTask(this.config.task)
             }
 
             const editTimeToFirstTokenSpan = tracer.startSpan('cody.edit.provider.timeToFirstToken')
@@ -240,7 +239,7 @@ export class EditProvider {
                 if (this.config.task.destinationFile) {
                     // We have already provided a destination file,
                     // Treat this as the test file to insert to
-                    await this.config.controller.didReceiveNewFileRequest(
+                    await this.config.fixupController.didReceiveNewFileRequest(
                         this.config.task.id,
                         this.config.task.destinationFile
                     )
@@ -284,7 +283,11 @@ export class EditProvider {
             if (modelsService.isStreamDisabled(model)) {
                 params.stream = false
             }
-            const stream = await this.config.chat.chat(messages, { ...params }, abortController.signal)
+            const stream = await this.config.chatClient.chat(
+                messages,
+                { ...params },
+                abortController.signal
+            )
 
             let textConsumed = 0
             let firstTokenReceived = false
@@ -414,11 +417,11 @@ export class EditProvider {
      * Will allow the user to view the error in more detail if needed.
      */
     protected handleError(error: Error): void {
-        this.config.controller.error(this.config.task.id, error)
+        this.config.fixupController.error(this.config.task.id, error)
     }
 
     private async handleFixupEdit(response: string, isMessageInProgress: boolean): Promise<void> {
-        return this.config.controller.didReceiveFixupText(
+        return this.config.fixupController.didReceiveFixupText(
             this.config.task.id,
             responseTransformer(response, this.config.task, isMessageInProgress),
             isMessageInProgress ? 'streaming' : 'complete'
@@ -426,7 +429,7 @@ export class EditProvider {
     }
 
     private async handleFixupInsert(response: string, isMessageInProgress: boolean): Promise<void> {
-        return this.config.controller.didReceiveFixupInsertion(
+        return this.config.fixupController.didReceiveFixupInsertion(
             this.config.task.id,
             responseTransformer(response, this.config.task, isMessageInProgress),
             isMessageInProgress ? 'streaming' : 'complete'
@@ -452,7 +455,7 @@ export class EditProvider {
             const newDoc = await workspace.openTextDocument({
                 language: currentDoc?.languageId,
             })
-            await this.config.controller.didReceiveNewFileRequest(this.config.task.id, newDoc.uri)
+            await this.config.fixupController.didReceiveNewFileRequest(this.config.task.id, newDoc.uri)
             return
         }
 
@@ -484,7 +487,7 @@ export class EditProvider {
                 newFileUri = newFileUri.with({ scheme: 'untitled' })
             }
             try {
-                await this.config.controller.didReceiveNewFileRequest(task.id, newFileUri)
+                await this.config.fixupController.didReceiveNewFileRequest(task.id, newFileUri)
             } catch (error) {
                 this.handleError(new Error('Cody failed to generate unit tests', { cause: error }))
             }
