+++ vscode/src/non-stop/FixupController.ts
@@ -2,8 +2,10 @@ import * as vscode from 'vscode'
 
 import {
     type ContextItem,
+    DEFAULT_EVENT_SOURCE,
     type EditModel,
     type EventSource,
+    EventSourceTelemetryMetadataMapping,
     type PromptString,
     type Rule,
     currentAuthStatus,
@@ -11,30 +13,26 @@ import {
     telemetryRecorder,
 } from '@sourcegraph/cody-shared'
 
+import type { SmartApplyResult } from '../chat/protocol'
+import { PersistenceTracker } from '../common/persistence-tracker'
+import { lines } from '../completions/text-processing'
 import { executeEdit } from '../edit/execute'
+import { type QuickPickInput, getInput } from '../edit/input/get-input'
 import {
     type EditIntent,
     EditIntentTelemetryMetadataMapping,
     type EditMode,
     EditModeTelemetryMetadataMapping,
 } from '../edit/types'
-import { logDebug } from '../output-channel-logger'
-import { splitSafeMetadata } from '../services/telemetry-v2'
-import { countCode } from '../services/utils/code-count'
-
-import {
-    DEFAULT_EVENT_SOURCE,
-    EventSourceTelemetryMetadataMapping,
-} from '@sourcegraph/cody-shared/src/chat/transcript/messages'
-import type { SmartApplyResult } from '../chat/protocol'
-import { PersistenceTracker } from '../common/persistence-tracker'
-import { lines } from '../completions/text-processing'
-import { type QuickPickInput, getInput } from '../edit/input/get-input'
 import { isStreamedIntent } from '../edit/utils/edit-intent'
 import { getOverriddenModelForIntent } from '../edit/utils/edit-models'
 import type { ExtensionClient } from '../extension-client'
 import { isRunningInsideAgent } from '../jsonrpc/isRunningInsideAgent'
+import { logDebug } from '../output-channel-logger'
 import { charactersLogger } from '../services/CharactersLogger'
+import { splitSafeMetadata } from '../services/telemetry-v2'
+import { countCode } from '../services/utils/code-count'
+
 import { FixupDocumentEditObserver } from './FixupDocumentEditObserver'
 import type { FixupFile } from './FixupFile'
 import { FixupFileObserver } from './FixupFileObserver'
@@ -61,7 +59,6 @@ export interface CreateTaskOptions {
     insertionPoint?: vscode.Position
     telemetryMetadata?: FixupTelemetryMetadata
     taskId?: FixupTaskID
-    isPrefetch?: boolean
 }
 
 // This class acts as the factory for Fixup Tasks and handles communication between the Tree View and editor
@@ -524,7 +521,6 @@ export class FixupController
         insertionPoint,
         telemetryMetadata,
         taskId,
-        isPrefetch,
     }: CreateTaskOptions): FixupTask {
         const authStatus = currentAuthStatus()
         const overriddenModel = getOverriddenModelForIntent(intent, model, authStatus)
@@ -546,9 +542,6 @@ export class FixupController
             taskId
         )
         this.tasks.set(task.id, task)
-        if (!isPrefetch) {
-            this.decorator.didCreateTask(task)
-        }
 
         return task
     }
